Added more specific zero and one types (addresses #26)

+++ src/AdaBoost.jl
@@ -153,7 +153,7 @@ function learn(
             f_idx = feature_indices[j]
             # classifier error is the sum of image weights where the classifier is right
             # ε = sum(map(img_idx -> labels[img_idx] ≠ votes[img_idx, f_idx] ? weights[img_idx] : zero(Integer), 1:num_imgs))
-            ε = sum([labels[img_idx] ≠ votes[img_idx, f_idx] ? weights[img_idx] : zero(Integer) for img_idx in 1:num_imgs])
+            ε = sum([labels[img_idx] ≠ votes[img_idx, f_idx] ? weights[img_idx] : zero(Float64) for img_idx in 1:num_imgs])
             
             classification_errors[j] = ε
         end
